fix: avoid cross origin request on Mac safari
Safari on Mac does not use browser's cookie when calling
GM_xmlhttpRequest (see Tampermonkey/tampermonkey#786).

When user is not logged in, Nanyangpt will redirect to non-https
log-in page, which causes a cross origin request and user will get
a warnning.

For now, exclude tampermonkey on Mac to make the request. If
other browsers are affected, exclude them also.